+++ src/screens/transaction_details/components/instructions/components/instruction/components/index.ts
@@ -4,6 +4,7 @@ import InnerInstruction from './inner_instruction';
 
 // defined instruction types
 import CreateAccount from './system/create_account';
+import Transfer from './system/transfer';
 
 export {
   Json,
@@ -11,6 +12,8 @@ export {
   InnerInstruction,
 };
 
+// system
 export {
   CreateAccount,
+  Transfer,
 };